backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.3",
    "alpaca-py>=0.43.2",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "joblib>=1.5.3",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "pytz>=2025.2",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
    "websockets>=16.0",
]

[tool.ruff]
# Allow lines to be as long as 120.
line-length = 120